Extract basic components into shared monorepo package
