+++ editor/package.json
@@ -20,6 +20,7 @@
     "@tanstack/react-query": "^5.20.2",
     "bcrypt": "^5.1.1",
     "clsx": "^2.1.0",
+    "component-library": "^1.0.0",
     "flexsearch": "^0.7.43",
     "format-quantity": "^3.0.0",
     "fraction.js": "^4.3.7",
