+++ editor/src/app/(recipes)/search/searchForm.tsx
@@ -4,8 +4,8 @@ import {
   MassagedRecipeEntry,
   ReadRecipeIndexResult,
 } from "recipes-collection/controller/data/readIndex";
-import { Button } from "@/components/Button";
-import { TextInput } from "@/components/Form/inputs/Text";
+import { Button } from "component-library/Button";
+import { TextInput } from "component-library/Form/inputs/Text";
 import {
   QueryClient,
   QueryClientProvider,
