+++ pnpm-workspace.yaml
@@ -2,3 +2,4 @@ packages:
   - 'website'
   - 'editor'
   - 'recipes-collection'
+  - 'component-library'
