+++ recipes-collection/components/Form/Ingredients/index.tsx
@@ -1,10 +1,13 @@
 import { RecipeFormErrors } from "../../../controller/formState";
 import { Ingredient } from "../../../controller/types";
 import { createIngredients } from "../../../util/parseIngredients";
-import { Button } from "@/components/Button";
-import { FieldWrapper, baseInputStyle } from "@/components/Form";
-import { InputListControls, useKeyList } from "@/components/Form/inputs/List";
-import { TextInput } from "@/components/Form/inputs/Text";
+import { Button } from "component-library/Button";
+import { FieldWrapper, baseInputStyle } from "component-library/Form";
+import {
+  InputListControls,
+  useKeyList,
+} from "component-library/Form/inputs/List";
+import { TextInput } from "component-library/Form/inputs/Text";
 import clsx from "clsx";
 import { useRef } from "react";
