+++ recipes-collection/components/View/Instructions/index.tsx
@@ -2,7 +2,7 @@
 
 import { InstructionEntry } from "../../../controller/types";
 
-import Markdown from "@/components/Markdown";
+import Markdown from "component-library/Markdown";
 
 const stepHeadingStyle = "text-lg font-bold my-2 border-b border-white";
 const childHeadingStyle = "text-base font-bold my-1 border-b border-white";
