+++ website/package.json
@@ -19,6 +19,7 @@
     "@sindresorhus/slugify": "^2.2.1",
     "@tanstack/react-query": "^5.20.2",
     "clsx": "^2.1.0",
+    "component-library": "^1.0.0",
     "flexsearch": "^0.7.43",
     "format-quantity": "^3.0.0",
     "fraction.js": "^4.3.7",
