--- website/src/components/Form/inputs/Date/index.tsx
@@ -1,36 +1,0 @@
-import { ChangeEventHandler } from "react";
-import { Errors, FieldWrapper, baseInputStyle } from "../..";
-import clsx from "clsx";
-
-export function DateInput({
-  name,
-  id = name,
-  defaultValue,
-  onChange,
-  label,
-  placeholder,
-  errors,
-}: {
-  name: string;
-  id?: string;
-  label: string;
-  defaultValue?: string;
-  onChange?: ChangeEventHandler<HTMLInputElement>;
-  placeholder?: string;
-  errors?: string[];
-}) {
-  return (
-    <FieldWrapper label={label} id={id}>
-      <Errors errors={errors} />
-      <input
-        type="date"
-        name={name}
-        id={id}
-        className={clsx(baseInputStyle, "px-2 py-1")}
-        defaultValue={defaultValue}
-        onChange={onChange}
-        placeholder={placeholder}
-      />
-    </FieldWrapper>
-  );
-}
