--- website/src/components/Form/inputs/Text/index.tsx
@@ -1,39 +1,0 @@
-import { ChangeEventHandler } from "react";
-import { Errors, FieldWrapper, baseInputStyle } from "../..";
-import clsx from "clsx";
-
-export function TextInput({
-  name,
-  id = name,
-  defaultValue,
-  onChange,
-  label,
-  placeholder,
-  errors,
-  list,
-}: {
-  name: string;
-  id?: string;
-  label?: string;
-  defaultValue?: string;
-  onChange?: ChangeEventHandler<HTMLInputElement>;
-  placeholder?: string;
-  errors?: string[];
-  list?: string;
-}) {
-  return (
-    <FieldWrapper label={label} id={id}>
-      <Errors errors={errors} />
-      <input
-        type="text"
-        name={name}
-        id={id}
-        className={clsx(baseInputStyle, "px-2 py-1")}
-        defaultValue={defaultValue}
-        onChange={onChange}
-        placeholder={placeholder}
-        list={list}
-      />
-    </FieldWrapper>
-  );
-}
