--- website/src/components/Markdown/styles.module.css
@@ -1,33 +1,0 @@
-.content {
-  h1,
-  h2,
-  h3,
-  h4,
-  h5,
-  h6 {
-    @apply font-bold;
-  }
-  h1 {
-    @apply text-3xl;
-  }
-  h2 {
-    @apply text-2xl;
-  }
-  h3 {
-    @apply text-xl;
-  }
-  h4 {
-    @apply text-lg;
-  }
-  h5 {
-  }
-  h6 {
-  }
-  ul {
-    list-style: disc;
-    padding-left: 1rem;
-  }
-  a {
-    text-decoration: underline;
-  }
-}
